requirements.txt
simpy>=4.1
PyYAML>=6.0
matplotlib>=3.8